+++ src/main/java/com/simibubi/create/content/equipment/blueprint/BlueprintItem.java
@@ -4,7 +4,6 @@ import java.util.Collection;
 
 import com.simibubi.create.AllItems;
 import com.simibubi.create.content.logistics.filter.AttributeFilterMenu.WhitelistMode;
-import com.simibubi.create.content.logistics.filter.FilterItem;
 import com.simibubi.create.content.logistics.item.filter.attribute.ItemAttribute;
 import com.simibubi.create.content.logistics.item.filter.attribute.attributes.InTagAttribute;
 
@@ -93,7 +92,7 @@ public class BlueprintItem extends Item {
 	}
 
 	private static ItemStack convertIngredientToFilter(Ingredient ingredient) {
-		Ingredient.Value[] acceptedItems = ingredient.values;
+		Value[] acceptedItems = ingredient.values;
 		if (acceptedItems == null || acceptedItems.length > 18)
 			return ItemStack.EMPTY;
 		if (acceptedItems.length == 0)
@@ -102,7 +101,7 @@ public class BlueprintItem extends Item {
 			return convertIItemListToFilter(acceptedItems[0]);
 
 		ItemStack result = AllItems.FILTER.asStack();
-		ItemStackHandler filterItems = FilterItem.getFilterItems(result);
+		ItemStackHandler filterItems = AllItems.FILTER.get().getFilterItemHandler(result);
 		for (int i = 0; i < acceptedItems.length; i++)
 			filterItems.setStackInSlot(i, convertIItemListToFilter(acceptedItems[i]));
 		result.getOrCreateTag()
@@ -133,7 +132,7 @@ public class BlueprintItem extends Item {
 
 		if (itemList instanceof MultiItemValue) {
 			ItemStack result = AllItems.FILTER.asStack();
-			ItemStackHandler filterItems = FilterItem.getFilterItems(result);
+			ItemStackHandler filterItems = AllItems.FILTER.get().getFilterItemHandler(result);
 			int i = 0;
 			for (ItemStack itemStack : stacks) {
 				if (i >= 18)
