+++ src/main/java/com/simibubi/create/content/equipment/blueprint/BlueprintOverlayRenderer.java
@@ -12,18 +12,14 @@ import com.simibubi.create.AllItems;
 import com.simibubi.create.content.equipment.blueprint.BlueprintEntity.BlueprintCraftingInventory;
 import com.simibubi.create.content.equipment.blueprint.BlueprintEntity.BlueprintSection;
 import com.simibubi.create.content.logistics.BigItemStack;
-import com.simibubi.create.content.logistics.filter.AttributeFilterMenu.WhitelistMode;
 import com.simibubi.create.content.logistics.filter.FilterItem;
 import com.simibubi.create.content.logistics.filter.FilterItemStack;
-import com.simibubi.create.content.logistics.item.filter.attribute.ItemAttribute;
-import com.simibubi.create.content.logistics.item.filter.attribute.attributes.InTagAttribute;
 import com.simibubi.create.content.logistics.packager.InventorySummary;
 import com.simibubi.create.content.logistics.tableCloth.BlueprintOverlayShopContext;
 import com.simibubi.create.content.logistics.tableCloth.ShoppingListItem.ShoppingList;
 import com.simibubi.create.content.logistics.tableCloth.TableClothBlockEntity;
 import com.simibubi.create.content.trains.track.TrackPlacement.PlacementInfo;
 import com.simibubi.create.foundation.gui.AllGuiTextures;
-import com.simibubi.create.foundation.item.ItemHelper;
 
 import net.createmod.catnip.animation.AnimationTickHolder;
 import net.createmod.catnip.data.Couple;
@@ -35,7 +31,6 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.inventory.tooltip.TooltipRenderUtil;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.nbt.ListTag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.CraftingContainer;
@@ -54,9 +49,6 @@ import net.minecraftforge.client.gui.overlay.ForgeGui;
 import net.minecraftforge.client.gui.overlay.IGuiOverlay;
 import net.minecraftforge.items.ItemHandlerHelper;
 import net.minecraftforge.items.ItemStackHandler;
-import net.minecraftforge.registries.ForgeRegistries;
-import net.minecraftforge.registries.tags.ITag;
-import net.minecraftforge.registries.tags.ITagManager;
 
 public class BlueprintOverlayRenderer {
 
@@ -438,34 +430,8 @@ public class BlueprintOverlayRenderer {
 
 	private static ItemStack[] getItemsMatchingFilter(ItemStack filter) {
 		return cachedRenderedFilters.computeIfAbsent(filter, itemStack -> {
-			CompoundTag tag = itemStack.getOrCreateTag();
-
-			if (AllItems.FILTER.isIn(itemStack) && !tag.getBoolean("Blacklist")) {
-				ItemStackHandler filterItems = FilterItem.getFilterItems(itemStack);
-				return ItemHelper.getNonEmptyStacks(filterItems).toArray(ItemStack[]::new);
-			}
-
-			if (AllItems.ATTRIBUTE_FILTER.isIn(itemStack)) {
-				WhitelistMode whitelistMode = WhitelistMode.values()[tag.getInt("WhitelistMode")];
-				ListTag attributes = tag.getList("MatchedAttributes", net.minecraft.nbt.Tag.TAG_COMPOUND);
-				if (whitelistMode == WhitelistMode.WHITELIST_DISJ && attributes.size() == 1) {
-					ItemAttribute fromNBT = ItemAttribute.loadStatic((CompoundTag) attributes.get(0));
-					if (fromNBT instanceof InTagAttribute inTag) {
-						ITagManager<Item> tagManager = ForgeRegistries.ITEMS.tags();
-						if (tagManager.isKnownTagName(inTag.tag)) {
-							ITag<Item> taggedItems = tagManager.getTag(inTag.tag);
-							if (!taggedItems.isEmpty()) {
-								ItemStack[] stacks = new ItemStack[taggedItems.size()];
-								int i = 0;
-								for (Item item : taggedItems) {
-									stacks[i] = new ItemStack(item);
-									i++;
-								}
-								return stacks;
-							}
-						}
-					}
-				}
+			if (itemStack.getItem() instanceof FilterItem filterItem) {
+				return filterItem.getFilterItems(itemStack);
 			}
 
 			return new ItemStack[0];
